Address Rob's PR comments
